remove old normalizer
