+++ rholang/src/main/scala/coop/rchain/rholang/interpreter/normalizer/instances/Proc/PSimpleTypeInstance.scala
@@ -18,7 +18,7 @@ import coop.rchain.rholang.ast.rholang_mercury.Absyn.{
   SimpleTypeString,
   SimpleTypeUri
 }
-import coop.rchain.rholang.interpreter.compiler.{ProcVisitInputs, ProcVisitOutputs}
+import coop.rchain.rholang.interpreter.compiler.Visit._
 import coop.rchain.rholang.interpreter.normalizer.Normalizer
 
 trait PSimpleTypeInstance {
