+++ rholang/src/test/scala/coop/rchain/rholang/interpreter/compiler/normalizer/ProcMatcherSpec.scala
@@ -24,11 +24,10 @@ import coop.rchain.rholang.interpreter.compiler.{
   IndexMapChain,
   NameSort,
   ProcSort,
-  ProcVisitInputs,
-  ProcVisitOutputs,
   SourcePosition,
   VarSort
 }
+import coop.rchain.rholang.interpreter.compiler.Visit._
 import coop.rchain.rholang.interpreter.normalizer.Normalizer
 import monix.eval.Coeval
